replaced EulerProblem interface by a superclass (subclass of Action)

+++ src/courseprojectprimes/actions/eulerproblems/FibonacciArray.java
@@ -9,8 +9,8 @@ package courseprojectprimes.actions.eulerproblems;
  */
 
 public class FibonacciArray {
-    public int[] array;
-    public FibonacciArray(int[] array) {
+    public long[] array;
+    public FibonacciArray(long[] array) {
         this.array = array;
     }
     
@@ -18,14 +18,14 @@ public class FibonacciArray {
         return getLastElement() % 2 == 0;
     }
     
-    public int getLastElement() {
+    public long getLastElement() {
         return this.array[this.array.length - 1];
     }
     
     public void computeNextFib() {
         // shift elements of the list to the left, dropping the first one,
         // generate the next fibonacci number and add it as the third element
-        int nextFib = this.array[1] + this.array[2];
+        long nextFib = this.array[1] + this.array[2];
         this.array[0] = this.array[1];
         this.array[1] = this.array[2];
         this.array[2] = nextFib;
